% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_with_go_terms.R
\name{annotate_with_go_terms}
\alias{annotate_with_go_terms}
\title{Annotate genes with GO terms from parsed cluego file}
\usage{
annotate_with_go_terms(gene_df, go_df, output_folder_path, species_prefix)
}
\description{
Annotate genes with GO terms from parsed cluego file
}
