% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_common.R
\name{annotate_with_kegg_common_pathways}
\alias{annotate_with_kegg_common_pathways}
\title{Annotate genes with KEGG common pathways}
\usage{
annotate_with_kegg_common_pathways(
  gene_df,
  species_prefix,
  kegg_df,
  output_folder_path
)
}
\arguments{
\item{gene_df}{Data frame with input genes. Must contain \code{Gene_symbol}.}

\item{kegg_df}{Data frame with KEGG annotations, containing \code{genes}
and \code{pathway_id}.}
}
\value{
A list with elements:
\item{annotated_genes}{Genes successfully mapped to KEGG pathways}
\item{not_mapped}{Genes without KEGG annotation}
}
\description{
Annotate genes with KEGG common pathways
}
