% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_coverage.R
\name{check_annotation_coverage}
\alias{check_annotation_coverage}
\title{Save unmapped genes to file}
\usage{
check_annotation_coverage(annotated_genes, genes_list, output_path)
}
\description{
Save unmapped genes to file
}
