% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_genes_to_orthologs.R
\name{convert_genes_to_orthologs}
\alias{convert_genes_to_orthologs}
\title{Convert a gene list into orthologs using internal ortholog tables}
\usage{
convert_genes_to_orthologs(
  genes_list_path,
  from_species,
  to_species,
  output_folder_path = file.path(Sys.getenv("USERPROFILE"), "Desktop",
    "Networks_results")
)
}
\arguments{
\item{genes_list_path}{Path to a file with input genes.
Must contain either \code{Ensembl_gene_id} or \code{Gene_symbol}.}

\item{from_species}{Species of the input genes ("human", "mouse", "macaque", "zebrafish").}

\item{to_species}{Species to convert into ("human", "mouse", "macaque").}

\item{output_folder_path}{Folder where results will be saved.}
}
\description{
Convert a gene list into orthologs using internal ortholog tables
}
\details{
This function loads internal ortholog tables, filters out rows without valid
original/ortholog Ensembl IDs, keeps only the rows corresponding to the input
gene list, and constructs a ready-to-use table of ortholog Ensembl IDs and symbols.

Resulting tables are saved as TXT files in \code{output_folder_path}.
}
