% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Project.R
\name{create_project_folder}
\alias{create_project_folder}
\title{Create working directory with example config files}
\usage{
create_project_folder(
  path = "my_project",
  copy_example_configs = FALSE,
  create_empty_gene_list = FALSE
)
}
\arguments{
\item{path}{Path to create the config project folder. Defaults to "my_project".}

\item{copy_example_configs}{Logical; if TRUE, copies example config files into the project folder. Default is FALSE.}

\item{create_empty_gene_list}{Logical; if TRUE, creates an empty gene list file inside the project folder. Default is FALSE.}
}
\description{
Create working directory with example config files
}
