% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_config_evidence.R
\name{default_evidence_experimental_only}
\alias{default_evidence_experimental_only}
\title{Get default path to config_evidence.txt}
\usage{
default_evidence_experimental_only()
}
\value{
Path to internal config_evidence.txt
}
\description{
Get default path to config_evidence.txt
}
