% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_GO_type.R
\name{detect_go_type}
\alias{detect_go_type}
\title{Detect GO type from file name}
\usage{
detect_go_type(go_file_path)
}
\arguments{
\item{go_file_path}{Path to GO file.}
}
\value{
GO category prefix: BP, MF, CC, IP.
}
\description{
Detect GO type from file name
}
