% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_go_annotations.R
\name{filter_go_annotations}
\alias{filter_go_annotations}
\title{Filter GO annotations by level and evidence}
\usage{
filter_go_annotations(
  annotated_genes,
  config_evidence_path,
  level_from,
  level_to
)
}
\description{
Filter GO annotations by level and evidence
}
