% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_go_neighbor.R
\name{find_go_neighbor}
\alias{find_go_neighbor}
\title{Find the neighbors of the target gene in GO annotation}
\usage{
find_go_neighbor(
  gene_name,
  species_prefix,
  go_type,
  output_folder_path = file.path(Sys.getenv("USERPROFILE"), "Desktop",
    "Networks_results"),
  use_ortholog = "none",
  config_evidence_path = default_evidence_experimental_only(),
  level_from = 5,
  level_to = 20,
  label_type = "SYMBOL",
  layout_name = "layout_nicely",
  threshold = 5
)
}
\arguments{
\item{gene_name}{Path to gene list (CSV or TXT).}

\item{species_prefix}{Short prefix (e.g., "human", "mouse").}

\item{go_type}{GO category: BP, CC, MF, IP.}

\item{output_folder_path}{Output directory (default: "./results").}

\item{use_ortholog}{= "none"}

\item{config_evidence_path}{Optional path to config_evidence.txt. If missing, uses internal.}

\item{level_from}{GO tree level (default: 5).}

\item{level_to}{GO tree level (default: 20).}

\item{label_type}{Gene label type (default: "SYMBOL").}

\item{layout_name}{Graph layout (default: "layout_nicely").}

\item{threshold}{Threshold for edge filtering (default: 5).}
}
\value{
Data frame of filtered annotations (invisible).
}
\description{
Find the neighbors of the target gene in GO annotation
}
