% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_go_file_by_prefix.R
\name{get_go_file_by_prefix}
\alias{get_go_file_by_prefix}
\title{Locate internal GO annotation file}
\usage{
get_go_file_by_prefix(species_prefix, go_type)
}
\arguments{
\item{species_prefix}{Short species code: human, mouse, zebrafish, macaque}

\item{go_type}{GO category prefix: BP, MF, CC, IP}
}
\value{
Full system path to matching GO annotation file
}
\description{
Locate internal GO annotation file
}
