% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_utils.R
\name{infer_species}
\alias{infer_species}
\title{Infer species from Ensembl gene ID}
\usage{
infer_species(ensembl_id)
}
\arguments{
\item{ensembl_id}{A character string with Ensembl ID (e.g. ENSG...).}
}
\value{
A species prefix (human, mouse, etc.) or NA if not matched.
}
\description{
Infer species from Ensembl gene ID
}
