% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_friday.R
\name{is_friday}
\alias{is_friday}
\title{Is Friday?}
\usage{
is_friday(date)
}
\arguments{
\item{date}{A Date class object or a string in the form "yyyymmdd".}
}
\value{
A boolean vector the same length as \code{date}.
}
\description{
Check which elements in a vector of dates are Fridays.
}
\examples{
# TRUE
is_friday("2025-03-28")
"2025-03-28" |> lubridate::ymd() |> is_friday()

# FALSE
is_friday("2025-03-24")

# TRUE, FALSE
c("2025-03-28", "2025-03-24") |> is_friday()

}
