% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_annotation.R
\name{load_annotation}
\alias{load_annotation}
\title{Load internal annotation data}
\usage{
load_annotation(
  source_type,
  species_prefix,
  connection_type = NULL,
  go_type = NULL
)
}
\arguments{
\item{source_type}{"GO" or "KEGG"}

\item{species_prefix}{e.g. "human", "mouse", "macaca", "zebrafish"}

\item{connection_type}{For KEGG: "common_KEGG_pathway" or "relation_in_KEGG_pathway"}

\item{go_type}{For GO: "BP", "CC", "MF", or "IP"}
}
\value{
Annotation data.frame
}
\description{
Load internal annotation data
}
\keyword{internal}
