% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ortholog.R
\name{load_ortholog}
\alias{load_ortholog}
\title{Load ortholog mapping}
\usage{
load_ortholog(from_species, to_species)
}
\arguments{
\item{from_species}{Species of input genes ("human", "mouse", "macacaque").}

\item{to_species}{Target species for mapping ("human", "mouse", "macacaque").}
}
\value{
Data frame with ortholog mapping, e.g., Ensembl IDs.
}
\description{
Load ortholog mapping
}
