% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_genes.R
\name{prepare_genes}
\alias{prepare_genes}
\title{Prepare gene list from input file}
\usage{
prepare_genes(genes_list_path, species_prefix, output_folder_path)
}
\arguments{
\item{genes_list_path}{Path to gene list file.}

\item{species_prefix}{Expected species.}

\item{output_folder_path}{Path for outputs (e.g., duplicated symbols).}
}
\value{
A data frame with Ensembl gene IDs.
}
\description{
Prepare gene list from input file
}
