% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_main_config.R
\name{read_config_txt}
\alias{read_config_txt}
\title{Read main configuration from a tab-delimited file}
\usage{
read_config_txt(config_path)
}
\arguments{
\item{config_path}{Path to a main_config.txt file.}
}
\value{
A named list of configuration parameters.
}
\description{
Read main configuration from a tab-delimited file
}
