% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_go_annotations.R
\name{read_go_annotations}
\alias{read_go_annotations}
\title{Read and annotate GO data}
\usage{
read_go_annotations(go_file_path, species_prefix)
}
\arguments{
\item{go_file_path}{Path to GO annotation file.}

\item{species_prefix}{Species prefix for mapping.}
}
\value{
A data frame with GO + SYMBOL + ENSEMBL.
}
\description{
Read and annotate GO data
}
