% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relation_in_kegg_pathway.R
\name{relation_in_kegg_pathway}
\alias{relation_in_kegg_pathway}
\title{Build KEGG relation graph for a gene list}
\usage{
relation_in_kegg_pathway(
  genes_list_path,
  species_prefix,
  output_folder_path = "./results",
  use_ortholog = "none",
  label_type = "SYMBOL",
  layout_name = "layout_nicely",
  subtype_filter = "compound"
)
}
\arguments{
\item{genes_list_path}{Path to the input file with a gene list.}

\item{species_prefix}{Character string defining which internal}

\item{output_folder_path}{Directory for saving results. Created if absent.}

\item{use_ortholog}{= "none"}

\item{label_type}{Label mode for graph vertices. \code{"SYMBOL"}.}

\item{layout_name}{Name of the igraph layout function to use.}

\item{subtype_filter}{Vector of subtype names to exclude from relations.}
}
\description{
Build KEGG relation graph for a gene list
}
