% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages.R
\docType{data}
\name{error_messages}
\alias{error_messages}
\title{Predefined error messages for Networks package}
\format{
An object of class \code{list} of length 16.
}
\usage{
error_messages
}
\description{
This object contains all standard error messages used across functions
in the package. It is internal and not exported to the user.
}
\keyword{internal}
