% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gene_list.R
\name{read_input_file}
\alias{read_input_file}
\title{Read a gene list file in various formats}
\usage{
read_input_file(path)
}
\arguments{
\item{path}{Path to a gene list file (csv, txt, tsv, xls, xlsx).}
}
\value{
A data frame with gene identifiers.
}
\description{
Read a gene list file in various formats
}
