% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run the full GO-based gene interaction pipeline}
\usage{
run_pipeline(config_path)
}
\arguments{
\item{config_path}{Path to the main_config.txt}
}
\description{
Run the full GO-based gene interaction pipeline
}
